Refactor Get command implementation and add team member fetching functionality
