+++ Commander/vault/TeamListCommand.cs
@@ -195,26 +195,8 @@ namespace Commander
             string teamUid,
             Action<Severity, string> logger)
         {
-            try
-            {
-                var request = new global::Enterprise.GetTeamMemberRequest
-                {
-                    TeamUid = Google.Protobuf.ByteString.CopyFrom(teamUid.Base64UrlDecode())
-                };
-
-                var response = (global::Enterprise.GetTeamMemberResponse)await vault.Auth.ExecuteAuthRest(
-                    "vault/get_team_members",
-                    request,
-                    typeof(global::Enterprise.GetTeamMemberResponse));
-
-                var members = response.EnterpriseUser?.Select(u => u.Email).ToList() ?? new List<string>();
-                return members;
-            }
-            catch (Exception ex)
-            {
-                logger?.Invoke(Severity.Warning, $"Error fetching team members for {teamUid}: {ex.Message}");
-                return new List<string>();
-            }
+            var members = await TeamMemberExtensions.FetchTeamMembers(vault, teamUid, logger);
+            return members.Select(m => m.Email).Where(e => !string.IsNullOrEmpty(e)).ToList();
         }
 
         private static List<TeamListItem> SortTeams(List<TeamListItem> teams, string sortBy)
